Removed all fallible casts (except in tests)

+++ tests/integration.rs
@@ -8,7 +8,7 @@ mod integration {
     fn test_use_library() -> Result<()> {
         let mut trj = XTCTrajectory::open_read("tests/1l2y.xtc")?;
         let num_atoms = trj.get_num_atoms()?;
-        let mut frame = Frame::with_len(num_atoms as usize);
+        let mut frame = Frame::with_len(num_atoms);
 
         trj.read(&mut frame)?;
         trj.read(&mut frame)?;
@@ -21,7 +21,7 @@ mod integration {
         let trj = XTCTrajectory::open_read("tests/1l2y.xtc")?;
         let frames: Result<Vec<Rc<Frame>>> = trj.into_iter().collect();
         for (idx, frame) in frames?.iter().enumerate() {
-            assert_eq!(frame.step as usize, idx + 1);
+            assert_eq!(frame.step, idx + 1);
         }
         Ok(())
     }
